shoot action

+++ CompareNodes.cpp
@@ -19,8 +19,8 @@ bool CompareNodes::operator ( )(const Node* n1, const Node* n2)
 
 	int x1 = p1.GetX(), y1 = p1.GetY(), x2 = p2.GetX(), y2 = p2.GetY();
 	Maze &maze = Maze::getInstance();
-	int safty1 = maze.parts[y1][x1].getSaftyScore();
-	int safty2 = maze.parts[y2][x2].getSaftyScore();
+	double safty1 = maze.parts[y1][x1].getSaftyScore();
+	double safty2 = maze.parts[y2][x2].getSaftyScore();
 
 	return n1->GetF()+(1-safty1) > n2->GetF()+(1-safty2);
 }
